+++ mail/mail-common/src/lib.rs
@@ -36,6 +36,14 @@ use proton_action_queue::action::Id as ActionId;
 use proton_api_mail::services::proton::common::MessageId;
 use thiserror::Error;
 
+// Avoid breaking back compat.
+//
+// TODO: We should probably use a better name at some point for the clients like "protonSdk" or something
+// but that would be a breaking change
+// (fixed with search and replace but something we need to coordinate.)
+#[cfg(feature = "uniffi")]
+uniffi::setup_scaffolding!();
+
 pub const ALL_LABEL_TYPES: [LabelType; 4] = [
     LabelType::Label,
     LabelType::ContactGroup,