+++ mail/mail-uniffi/src/lib.rs
@@ -162,7 +162,7 @@ pub mod errors;
 mod log;
 pub mod mail;
 
-uniffi::setup_scaffolding!();
+uniffi::setup_scaffolding!("proton_mail_uniffi");
 
 /// A callback interface for live queries.
 ///
@@ -184,7 +184,6 @@ pub trait LiveQueryCallback: Send + Sync {
 ///
 /// This handle can be used to disconnect from the live query.
 ///
-#[allow(dead_code)]
 #[derive(uniffi::Object)]
 pub struct WatchHandle(RealWatchHandle);